Implement CalculateEDSMDistance. Clean up TriggerSystemChange (Needs MORE cleaning) Fixes #11

+++ RatTracker-WPF/RatTracker-WPF/Models/EDSMSystem.cs
@@ -13,8 +13,8 @@ namespace RatTracker_WPF.Models
     }
     public class EDSMCoords
     {
-        public float x { get; set; }
-        public float y { get; set; }
-        public float z { get; set; }
+        public double x { get; set; }
+        public double y { get; set; }
+        public double z { get; set; }
     }
 }
